cp encoding refactor
