--- src/data_preparation/encodings/CP.py
@@ -1,71 +1,0 @@
-import pickle
-from tqdm import tqdm
-import numpy as np
-
-import src.data_preparation.utils as utils
-
-class CP(object):
-    def __init__(self, dict):
-        self.event2word, self.word2event = pickle.load(open(dict, 'rb'))
-        self.pad_word = [self.event2word[etype]['%s <PAD>' % etype] for etype in self.event2word]
-
-    def extract_events(self, input_path):
-        note_items, tempo_items = utils.read_items(input_path)
-        if len(note_items) == 0:  # if the midi contains nothing
-            return None
-        note_items = utils.quantize_items(note_items)
-        max_time = note_items[-1].end
-        items = tempo_items + note_items
-
-        groups = utils.group_items(items, max_time)
-        events = utils.item2event(groups)
-        return events
-
-    def padding(self, data, max_len, ans):
-        pad_len = max_len - len(data)
-        for _ in range(pad_len):
-            if not ans:
-                data.append(self.pad_word)
-            else:
-                data.append(0)
-
-        return data
-
-    def prepare_data(self, midi_paths, max_len):
-        all_words, all_ys = [], []
-
-        for path in tqdm(midi_paths):
-            # extract events
-            events = self.extract_events(path)
-            if not events:  # if midi contains nothing
-                print(f'skip {path} because it is empty')
-                continue
-
-            # events to words
-            words, ys = [], []
-            for note_tuple in events:
-                nts, to_class = [], -1
-                for e in note_tuple:
-                    e_text = '{} {}'.format(e.name, e.value)
-                    nts.append(self.event2word[e.name][e_text])
-                    if e.name == 'Pitch':
-                        to_class = e.event_type
-                words.append(nts)
-
-            # slice to chunks so that max length = max_len (default: 512)
-            slice_words, slice_ys = [], []
-            for i in range(0, len(words), max_len):
-                slice_words.append(words[i:i + max_len])
-                slice_ys.append(ys[i:i + max_len])
-
-            # padding or drop
-            if len(slice_words[-1]) < max_len:
-                slice_words[-1] = self.padding(slice_words[-1], max_len, ans=False)
-
-            all_words = all_words + slice_words
-            all_ys = all_ys + slice_ys
-
-        all_words = np.array(all_words)
-        all_ys = np.array(all_ys)
-
-        return all_words, all_ys
